updates per reviewer requests
